requirements.txt
Flask==1.0.2
google-cloud-storage==1.14.0
numpy==1.16.2
